Model training code is working, changes to data module in progress
Trying to support a standardized model training workflow, with easy
generation of necessary pickle files and utilities for producing data
ready for inference. So far I have tested that the data object (C3SD)
is working and I can get a model trained up. Need to test the other changes
that are more oriented toward doing inference with a trained model